Separate GearElement into new module
