+++ ross/__init__.py
@@ -6,6 +6,7 @@ import ross.plotly_theme
 from .bearing_seal_element import *
 from .faults import *
 from .disk_element import *
+from .gear_element import *
 from .materials import *
 from .point_mass import *
 from .probe import *
