+++ ross/multi_rotor.py
@@ -1,226 +1,11 @@
 import numpy as np
 from re import search
 from copy import deepcopy as copy
-from plotly import graph_objects as go
 
-from ross.disk_element import DiskElement6DoF
+from ross.gear_element import GearElement
 from ross.rotor_assembly import Rotor
-from ross.units import Q_
 
-__all__ = ["GearElement", "MultiRotor"]
-
-
-class GearElement(DiskElement6DoF):
-    """A gear element.
-
-    This class creates a gear element from input data of inertia and mass.
-
-    Parameters
-    ----------
-    n: int
-        Node in which the gear will be inserted.
-    m : float
-        Mass of the gear element.
-    Id : float
-        Diametral moment of inertia.
-    Ip : float
-        Polar moment of inertia
-    pitch_diameter : float
-        Pitch diameter of the gear.
-    pressure_angle : float, optional
-        The pressure angle of the gear in radians. Default is 20° (converted to radians).
-    tag : str, optional
-        A tag to name the element.
-        Default is None.
-    scale_factor: float or str, optional
-        The scale factor is used to scale the gear drawing.
-        For gears it is also possible to provide 'mass' as the scale factor.
-        In this case the code will calculate scale factors for each gear based
-        on the gear with the higher mass. Notice that in this case you have to
-        create all gears with the scale_factor='mass'.
-        Default is 1.
-    color : str, optional
-        A color to be used when the element is represented.
-        Default is 'Goldenrod'.
-
-    Examples
-    --------
-    """
-
-    def __init__(
-        self,
-        n,
-        m,
-        Id,
-        Ip,
-        pitch_diameter,
-        pressure_angle=Q_(20, "deg"),
-        tag=None,
-        scale_factor=1.0,
-        color="Goldenrod",
-    ):
-
-        self.base_radius = pitch_diameter * np.cos(pressure_angle) / 2
-        self.pressure_angle = pressure_angle
-
-        super().__init__(n, m, Id, Ip, tag, scale_factor, color)
-
-    @classmethod
-    def from_geometry(
-        cls,
-        n,
-        material,
-        width,
-        i_d,
-        o_d,
-        pressure_angle=Q_(20, "deg"),
-        tag=None,
-        scale_factor=1.0,
-        color="Goldenrod",
-    ):
-        """Create a gear element from geometry properties.
-
-        This class method will create a gear element from geometry data.
-        Properties are calculated as per :cite:`friswell2010dynamics`, appendix 1
-        for a hollow cylinder:
-
-        Mass:
-
-        :math:`m = \\rho \\pi w (d_o^2 - d_i^2) / 4`
-
-        Polar moment of inertia:
-
-        :math:`I_p = m (d_o^2 + d_i^2) / 8`
-
-        Diametral moment of inertia:
-
-        :math:`I_d = \\frac{1}{2} I_p + \\frac{1}{12} m w^2`
-
-        Where :math:`\\rho` is the material density, :math:`w` is the gear width,
-        :math:`d_o` is the outer diameter and :math:`d_i` is the inner diameter.
-
-        Parameters
-        ----------
-        n : int
-            Node in which the gear will be inserted.
-        material: ross.Material
-            Gear material.
-        width : float
-            The face width of the gear (considering that the gear body has the same thickness).
-        i_d : float
-            Inner diameter (the diameter of the shaft on which the gear is mounted).
-        o_d : float
-            Outer pitch diameter.
-        pressure_angle : float, optional
-            The pressure angle of the gear in radians. Default is 20° (converted to radians).
-        tag : str, optional
-            A tag to name the element
-            Default is None
-        scale_factor: float, optional
-            The scale factor is used to scale the disk drawing.
-            Default is 1.
-        color : str, optional
-            A color to be used when the element is represented.
-            Default is 'Goldenrod'.
-
-        Attributes
-        ----------
-        m : float
-            Mass of the gear element.
-        Id : float
-            Diametral moment of inertia.
-        Ip : float
-            Polar moment of inertia
-
-        Examples
-        --------
-        """
-        m = material.rho * np.pi * width * (o_d**2 - i_d**2) / 4
-        Ip = m * (o_d**2 + i_d**2) / 8
-        Id = 1 / 2 * Ip + 1 / 12 * m * width**2
-
-        return cls(n, m, Id, Ip, o_d, pressure_angle, tag, scale_factor, color)
-
-    def _patch(self, position, fig):
-        """Gear element patch.
-
-        Patch that will be used to draw the gear element using plotly library.
-
-        Parameters
-        ----------
-        position : float
-            Position in which the patch will be drawn.
-        fig : plotly.graph_objects.Figure
-            The figure object which traces are added on.
-
-        Returns
-        -------
-        fig : plotly.graph_objects.Figure
-            The figure object which traces are added on.
-        """
-
-        zpos, ypos, yc_pos, scale_factor = position
-        scale_factor *= 1.3
-        radius = self.base_radius * 1.1 + 0.05
-
-        z_upper = [
-            zpos + scale_factor / 25,
-            zpos + scale_factor / 25,
-            zpos - scale_factor / 25,
-            zpos - scale_factor / 25,
-        ]
-        y_upper = [ypos, ypos + radius, ypos + radius, ypos]
-
-        z_lower = [
-            zpos + scale_factor / 25,
-            zpos + scale_factor / 25,
-            zpos - scale_factor / 25,
-            zpos - scale_factor / 25,
-        ]
-        y_lower = [-ypos, -ypos - radius, -ypos - radius, -ypos]
-
-        z_pos = z_upper
-        z_pos.append(None)
-        z_pos.extend(z_lower)
-
-        y_pos = y_upper
-        y_upper.append(None)
-        y_pos.extend(y_lower)
-
-        customdata = [self.n, self.Ip, self.Id, self.m, self.base_radius * 2]
-        hovertemplate = (
-            f"Gear Node: {customdata[0]}<br>"
-            + f"Polar Inertia: {customdata[1]:.3e}<br>"
-            + f"Diametral Inertia: {customdata[2]:.3e}<br>"
-            + f"Gear Mass: {customdata[3]:.3f}<br>"
-            + f"Gear Base Diam.: {customdata[4]:.3f}<br>"
-        )
-
-        fig.add_trace(
-            go.Scatter(
-                x=z_pos,
-                y=[y + yc_pos if y is not None else None for y in y_pos],
-                customdata=[customdata] * len(z_pos),
-                text=hovertemplate,
-                mode="lines",
-                fill="toself",
-                fillcolor=self.color,
-                fillpattern=dict(
-                    shape="/", fgcolor="rgba(0, 0, 0, 0.2)", bgcolor=self.color
-                ),
-                opacity=0.8,
-                line=dict(width=2.0, color="rgba(0, 0, 0, 0.2)"),
-                showlegend=False,
-                name=self.tag,
-                legendgroup="gears",
-                hoveron="points+fills",
-                hoverinfo="text",
-                hovertemplate=hovertemplate,
-                hoverlabel=dict(bgcolor=self.color),
-            )
-        )
-
-        return fig
+__all__ = ["MultiRotor"]
 
 
 class MultiRotor(Rotor):
@@ -270,7 +55,6 @@ class MultiRotor(Rotor):
         position="above",
         tag=None,
     ):
-
         self.rotors = [driving_rotor, driven_rotor]
         self.gear_ratio = gear_ratio
         self.gear_mesh_stiffness = gear_mesh_stiffness
